// Frontend/src/app/interfaces/entity.interface.ts
export interface Entity {
    id_persona: number;
    nombre: string;
    apellido: string;
    edad: number;
    dni: string;
    telefono: number;
    email: string;
    direccion: string;
    mascota: string;
    tipoInmueble: string;
    descripcion: string;
    isActive: boolean;
    status: string;
}